requirements.txt
alembic==1.8.1
bcrypt==4.0.0
blinker==1.5
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
protobuf==3.20.1
psycopg2==2.9.3
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.21.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
